update: リファクタリング
