+++ backend/src/openapi_server/ai/furniture.py
@@ -1,53 +1,17 @@
 import openai
-import json
-import base64
-from pathlib import Path
 
 from fastapi import UploadFile
 
-# OpenAI APIキーをロード
-def load_api_key():
-    file_path = Path(__file__).parent / 'secret.json'
-    with open(file_path, 'r') as file:
-        secrets = json.load(file)
-    return secrets['OPENAI_API_KEY']
-
-# 製品情報データをロード
-def load_product_data():
-    file_path = Path(__file__).parent / 'product_data.json'
-    with open(file_path, 'r') as file:
-        product_data = json.load(file)
-    return product_data
-
-# 画像をBase64でエンコード
-async def encode_image_to_base64(image: UploadFile):
-    image_data = await image.read()
-    return base64.b64encode(image_data).decode('utf-8')
-
-# 製品情報データから色のインデックスを取得
-def select_color_index(color: str):
-    color_list = load_product_data()['colors']
-    for i, c in enumerate(color_list):
-        if color in c or c in color:
-            return i
-    raise -1
-
-# 製品情報データからカテゴリのインデックスを取得
-def select_category_index(category: str):
-    category_list = load_product_data()['categories']
-    for i, c in enumerate(category_list):
-        if category in c or c in category:
-            return i
-    raise -1
+from openapi_server.ai import util
 
 class FurnitureDescribe:
     def __init__(self):
-        openai.api_key = load_api_key()
-        self.product_data = load_product_data()
+        openai.api_key = util.load_api_key()
+        self.product_data = util.load_product_data()
     
     async def get_describe(self, image: UploadFile):
         # 画像をBase64エンコード
-        base64_image = await encode_image_to_base64(image)
+        base64_image = await util.encode_image_to_base64(image)
         # OpenAI APIリクエストの実行
         response = openai.ChatCompletion.create(
             model="gpt-4o",
@@ -83,8 +47,8 @@ class FurnitureDescribe:
             max_tokens=300,
         )
         name, description, color, category = response.choices[0].message["content"].split(",")
-        color_idx = select_color_index(color)
-        category_idx = select_category_index(category)
+        color_idx = util.select_color_index(color)
+        category_idx = util.select_category_index(category)
         return {
             "product_name"  : name, 
             "description"   : description, 
@@ -94,14 +58,14 @@ class FurnitureDescribe:
 
 class FurnitureRecommendation:
     def __init__(self):
-        openai.api_key = load_api_key()
-        self.product_data = load_product_data()
+        openai.api_key = util.load_api_key()
+        self.product_data = util.load_product_data()
 
     async def get_recommend_color(self, image: UploadFile):
         color_list = self.product_data['colors']
 
         # 画像をBase64エンコード
-        base64_image = await encode_image_to_base64(image)
+        base64_image = await util.encode_image_to_base64(image)
         # OpenAI APIリクエストの実行
         response = openai.ChatCompletion.create(
             model="gpt-4o",
@@ -129,7 +93,7 @@ class FurnitureRecommendation:
             max_tokens=300,
         )
         color, reason = response.choices[0].message["content"].split(",")
-        color_index = select_color_index(color)
+        color_index = util.select_color_index(color)
         return {
             "color"     : color_index,
             "reason"    : reason
